Remove b flag to fopen.
fopen manpage says:

The mode string can also include the letter  ``b''  either  as  a  last
character  or  as a character between the characters in any of the two-
haracter strings described above.  This is strictly for  compatibility
with  ANSI  X3.159-1989  (``ANSI  C'')  and has no effect; the ``b'' is
ignored on all POSIX conforming systems, including Linux.  (Other  sys-
tems  may treat text files and binary files differently, and adding the
b'' may be a good idea if you do I/O to a binary file and expect that
your program may be ported to non-Unix environments.)

Removing, as POSIX specifies that this is ignored.